requirements.txt
streamlit
torch
torchvision
plotly
qiskit
qiskit-machine-learning
qiskit-aer
scikit-learn
requests
pillow
python-dotenv
numpy<2.0
